Refactor PDF generation to its own class
